requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
Pillow
python-dotenv>